Refactor client monitoring API to use service
Also made maximum VFS directory size configurable.

Fixes: #2005
