+++ config/proto/config.proto
@@ -856,6 +856,8 @@ message Defaults {
     // changed. This causes a lot of load on large deployments so it
     // is off by default.
     bool event_change_notify_all_clients = 6;
+
+    uint64 max_vfs_directory_size = 7;
 }
 
 // Configures crypto preferences
