+++ gui/velociraptor/src/components/sidebar/navigator.js
@@ -1,5 +1,6 @@
 import "./navigator.css";
 import logo from  "./velo.svg";
+import UserConfig from '../core/user.js';
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
@@ -11,6 +12,8 @@ import T from '../i8n/i8n.js';
 import { EncodePathInURL } from '../utils/paths.js';
 
 class VeloNavigator extends Component {
+    static contextType = UserConfig;
+
     static propTypes = {
         client: PropTypes.object.isRequired,
         vfs_path: PropTypes.string,
@@ -49,6 +52,9 @@ class VeloNavigator extends Component {
         if (this.props.vfs_path.length) {
             vfs_path = this.props.vfs_path;
         }
+
+        let customization = this.context.traits && this.context.traits.customizations;
+        customization = customization || {};
 
         return (
             <>
@@ -101,17 +107,19 @@ class VeloNavigator extends Component {
                         </ul>
                       </NavLink>
 
-                      <NavLink to="/events/server">
-                        <ul className="nav nav-pills  navigator">
-                          <li className="nav-link" state="server_events" >
-                            <span>
-                              <i className="navicon">
-                                <FontAwesomeIcon icon="eye"/></i>
-                            </span>
-                            {T("Server Events")}
-                          </li>
-                        </ul>
-                      </NavLink>
+                      { !customization.disable_server_events &&
+                        <NavLink to="/events/server">
+                          <ul className="nav nav-pills  navigator">
+                            <li className="nav-link" state="server_events" >
+                              <span>
+                                <i className="navicon">
+                                  <FontAwesomeIcon icon="eye"/></i>
+                              </span>
+                              {T("Server Events")}
+                            </li>
+                          </ul>
+                        </NavLink>
+                      }
 
                       <NavLink to="/collected/server">
                         <ul className="nav nav-pills  navigator">
